+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+# 2.0.0
+* Angel 2 support.
+
 # 1.0.0+1
 * Dart2 updates + backwards compatibility assurance.